refactor: centralize CLI agent launch validation
