+++ packages/cli/src/commands/workflow.ts
@@ -18,10 +18,9 @@ import {
   buildHeadlessRunContext,
   resolveCliRunModel,
 } from '../runtime/runModel';
-import { resolveCliAgent } from '../runtime/agents';
+import { assertCliAgentLaunch, resolveCliAgent } from '../runtime/agents';
 import { initLocalCliPlatform } from '../runtime/initPlatform';
 
-import { assertCliAgentLaunch } from './_helpers/agentLookupText';
 import { defineCliCommand } from './_helpers/defineCliCommand';
 import { emitCliResult } from './_helpers/output';
 import {
