+++ src/test-kernel/cli/AgentsRunCommand.vitest.mts
@@ -9,6 +9,7 @@ const mocks = vi.hoisted(() => {
     executeCliRequest: vi.fn(),
     withExpandedRunInputs: vi.fn(),
     initLocalCliPlatform: vi.fn(),
+    isAuthenticated: vi.fn(),
     resolveCliAgent: vi.fn(),
     resolveCliRunModel: vi.fn(),
     writeErrorStderr: vi.fn(),
@@ -17,6 +18,9 @@ const mocks = vi.hoisted(() => {
 });
 
 vi.mock('@agent/index', () => ({
+  getAgent: vi.fn(),
+  getAgentsByCategory: vi.fn(),
+  getVisibleAgents: vi.fn(),
   loadAgents: vi.fn(),
 }));
 
@@ -26,6 +30,12 @@ vi.mock('@agent/storage', () => ({
 
 vi.mock('@cli/runtime/initPlatform', () => ({
   initLocalCliPlatform: mocks.initLocalCliPlatform,
+}));
+
+vi.mock('@cli/runtime/supabaseAuth', () => ({
+  getCliAuthProvider: () => ({
+    isAuthenticated: mocks.isAuthenticated,
+  }),
 }));
 
 vi.mock('@cli/runtime/runModel', () => ({
@@ -47,7 +57,8 @@ vi.mock('@cli/commands/_helpers/output', () => ({
   emitCliResult: vi.fn(),
 }));
 
-vi.mock('@cli/runtime/agents', () => ({
+vi.mock('@cli/runtime/agents', async (importOriginal) => ({
+  ...(await importOriginal<typeof import('@cli/runtime/agents')>()),
   resolveCliAgent: mocks.resolveCliAgent,
 }));
